Add custom user model and create_superuser method
This commit adds a custom user model that extends Django's AbstractUser. The User model has a unique email field and a PhoneNumberField. The create_superuser method is also defined in the custom user model to ensure that superusers have the required permissions set. The USERNAME_FIELD and REQUIRED_FIELDS attributes are also set to enable the email and phone number fields to be used for authentication. The __str__ method is also implemented to return a string representation of the user.