Import get_user_model() function from Django auth
This commit imports the get_user_model() function from the django.contrib.auth module. The get_user_model() function is used to retrieve the currently active user model in a Django project, as specified by the AUTH_USER_MODEL setting. By importing this function, we can ensure that we always use the correct user model in our code, even if it is changed in the future.